ingestion_simulator/pyproject.toml
[project]
name = "ingestion-simulator"
version = "0.1.0"
description = "Data generator for Retail Analytics"
requires-python = ">=3.12"
dependencies = [
    "faker>=30.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "mysql-connector-python>=9.0.0",
    "click>=8.0.0",
]

[project.scripts]
ingest = "ingestion_simulator.cli:cli"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
